% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_folder.R
\name{populate_folder}
\alias{populate_folder}
\title{Autopopulates a New Analysis with Common Files}
\usage{
populate_folder(list_of_dirs = c("R", "data", "results"))
}
\arguments{
\item{list_of_dirs}{A character vector of
the names of folders to be created. Defaults
to c("R", "data", "results")}
}
\value{
Nothing explicit; this function is
  called for its side effects.
}
\description{
This function sets up the basic directory
structure I use in most of my analyses. By
default, this will produce folders for
R/, data/, and results/.
}
\examples{
populate_folder()
}
