% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisbid2018live.R
\docType{package}
\name{sisbid2018live}
\alias{sisbid2018live}
\alias{sisbid2018live-package}
\title{Basic Compilation of RR Scripts}
\description{
This package serves to collect scripts and
templates for producing reports in a more
reproducible fashion.
}
